+++ backend/migrations/007_audit_log_views.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS audit_log_views (
-  id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-  audit_log_id BIGINT UNSIGNED NOT NULL,
-  user_id BIGINT UNSIGNED NOT NULL,
+  id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
+  audit_log_id INT UNSIGNED NOT NULL,
+  user_id INT UNSIGNED NOT NULL,
   viewed_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   UNIQUE KEY uniq_audit_log_views_audit_user (audit_log_id, user_id),
   INDEX idx_audit_log_views_user_id (user_id),
